+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Context to share states with other apps
+- Export `PickupModal` and `LocationModal`
+
 ## [1.2.0] - 2025-06-06
 
 ### Added